fastapi-document-extraction/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pyodbc
pypdf2
python-docx
pillow
pytesseract
groq
python-dotenv
pydantic[email]
openpyxl